Fixed a bug with the ProcessUpdateObject method where it wasn't checking properly for lack of values.  This equates to Bug 169908, GitHub issue #35.

+++ DSLink/Request/SubscriptionUpdate.cs
@@ -11,9 +11,9 @@ namespace DSLink.Request
         public readonly DateTime Updated;
         public readonly string Status;
         public readonly int? Count;
-        public readonly int? Sum;
-        public readonly int? Min;
-        public readonly int? Max;
+        public readonly double? Sum;
+        public readonly double? Min;
+        public readonly double? Max;
 
         public SubscriptionUpdate(int subID, JToken value, string updated)
         {
@@ -29,7 +29,7 @@ namespace DSLink.Request
             Updated = DateTime.Parse(updated, null, DateTimeStyles.RoundtripKind);
         }
 
-        public SubscriptionUpdate(int subID, JToken value, string updated, int count, int sum, int min, int max)
+        public SubscriptionUpdate(int subID, JToken value, string updated, int count, double sum, double min, double max)
         {
             SubscriptionID = subID;
             Value = value;
